gpu: use static unsorted draw mode for distant zones
(cherry picked from commit ce01f6419c9cd732009749330df03d6ef6b2f1bf)

+++ runelite-client/src/main/java/net/runelite/client/plugins/gpu/SceneUploader.java
@@ -869,9 +869,9 @@ class SceneUploader
 
 	static void computeFaceUvs(Model model, int face)
 	{
-		final float[] vertexX = model.getVerticesX();
-		final float[] vertexY = model.getVerticesY();
-		final float[] vertexZ = model.getVerticesZ();
+		final int[] vertexX = model.getVerticesX();
+		final int[] vertexY = model.getVerticesY();
+		final int[] vertexZ = model.getVerticesZ();
 
 		final int[] indices1 = model.getFaceIndices1();
 		final int[] indices2 = model.getFaceIndices2();
